requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
requests
python-jose[cryptography]
passlib[bcrypt]
google-genai
pydantic
python-multipart